refactor: move withdrawCow to its own file
